+++ commons-http/src/test/java/com/hb0730/commons/http/support/httpclient/HttpClientAsyncImplTest.java
@@ -62,8 +62,8 @@ public class HttpClientAsyncImplTest {
     @Test
     public void testGetTest() {
         //有参请求
-        MapBuilder<String, String> build = MapBuilder.build();
-        Map<String, String> params = build.put("count", "2").builder();
+        MapBuilder<String, String> build = MapBuilder.builder();
+        Map<String, String> params = build.put("count", "2").build();
         HttpClientAsyncImpl async = new HttpClientAsyncImpl();
         async.get("http://poetry.apiopen.top/getTime", params, new HttpCallback() {
             @Override
@@ -81,8 +81,8 @@ public class HttpClientAsyncImplTest {
     @Test
     public void testGet1Test() {
         //多参请求
-        MapBuilder<String, String> build = MapBuilder.build();
-        Map<String, String> params = build.put("count", "2").builder();
+        MapBuilder<String, String> build = MapBuilder.builder();
+        Map<String, String> params = build.put("count", "2").build();
         //http://poetry.apiopen.top/poetryFull?page=1
         HttpClientAsyncImpl async = new HttpClientAsyncImpl();
         async.get("", params, new HttpCallback() {
@@ -101,12 +101,12 @@ public class HttpClientAsyncImplTest {
     @Test
     public void testGet2Test() {
         //请求头 参数判断
-        MapBuilder<String, String> build = MapBuilder.build();
+        MapBuilder<String, String> build = MapBuilder.builder();
         Map<String, String> params = build
                 .put("count", "2")
                 .put("page", "1")
                 .put("name", "李白")
-                .builder();
+                .build();
         HttpHeader httpHeader = HttpHeader.builder();
         httpHeader.add("Accept", "*/*");
         //测试为空,为null
@@ -131,12 +131,12 @@ public class HttpClientAsyncImplTest {
     @Test
     public void testGet3Test() {
         //请求头 参数判断
-        MapBuilder<String, String> build = MapBuilder.build();
+        MapBuilder<String, String> build = MapBuilder.builder();
         Map<String, String> params = build
                 .put("count", "2")
                 .put("page", "1")
                 .put("name", "李白")
-                .builder();
+                .build();
         HttpHeader httpHeader = HttpHeader.builder();
         httpHeader.add("Accept", "*/*");
         //测试为空,为null
@@ -196,13 +196,13 @@ public class HttpClientAsyncImplTest {
 
     @Test
     public void testPost1Test() {
-        MapBuilder<String, String> mapBuilder = MapBuilder.build();
+        MapBuilder<String, String> mapBuilder = MapBuilder.builder();
         mapBuilder.put("type", "all");
         mapBuilder.put("page", "1");
         mapBuilder.put("count", "10");
         // https://api.apiopen.top/getJoke
         HttpClientAsyncImpl async = new HttpClientAsyncImpl();
-        async.post("", mapBuilder.builder(), new HttpCallback() {
+        async.post("", mapBuilder.build(), new HttpCallback() {
             @Override
             public void success(String result) throws IOException {
                 Assert.assertNotNull("获取失败", result);
@@ -237,7 +237,7 @@ public class HttpClientAsyncImplTest {
 
     @Test
     public void testPost3Test() {
-        MapBuilder<String, String> mapBuilder = MapBuilder.build();
+        MapBuilder<String, String> mapBuilder = MapBuilder.builder();
         mapBuilder.put("type", "all");
         mapBuilder.put("page", "1");
         mapBuilder.put("count", "10");
@@ -245,7 +245,7 @@ public class HttpClientAsyncImplTest {
         HttpHeader header = HttpHeader.builder().add("Accept", "*/*");
         HttpClientAsyncImpl async = new HttpClientAsyncImpl();
         async.setHeader(header);
-        async.post("", mapBuilder.builder(), new HttpCallback() {
+        async.post("", mapBuilder.build(), new HttpCallback() {
             @Override
             public void success(String result) throws IOException {
                 Assert.assertNotNull("获取失败", result);