+++ commons-http/src/test/java/com/hb0730/commons/http/support/httpclient/HttpClientSyncImplTest.java
@@ -55,8 +55,8 @@ public class HttpClientSyncImplTest {
     public void testGetTest() {
         //有参请求
         HttpClientSyncImpl sync = new HttpClientSyncImpl();
-        MapBuilder<String, String> build = MapBuilder.build();
-        Map<String, String> params = build.put("count", "2").builder();
+        MapBuilder<String, String> build = MapBuilder.builder();
+        Map<String, String> params = build.put("count", "2").build();
         //http://poetry.apiopen.top/getTime
         String result = sync.get("", params);
         Assert.assertNotNull("获取失败", result);
@@ -66,8 +66,8 @@ public class HttpClientSyncImplTest {
     public void testGet2Test() {
         //多参请求
         HttpClientSyncImpl sync = new HttpClientSyncImpl();
-        MapBuilder<String, String> build = MapBuilder.build();
-        Map<String, String> params = build.put("count", "2").builder();
+        MapBuilder<String, String> build = MapBuilder.builder();
+        Map<String, String> params = build.put("count", "2").build();
         //http://poetry.apiopen.top/poetryFull?page=1
         String result = sync.get("", params);
         Assert.assertNotNull("获取失败", result);
@@ -78,12 +78,12 @@ public class HttpClientSyncImplTest {
     public void testGet1Test() {
         //请求头 参数判断
         HttpClientSyncImpl sync = new HttpClientSyncImpl();
-        MapBuilder<String, String> build = MapBuilder.build();
+        MapBuilder<String, String> build = MapBuilder.builder();
         Map<String, String> params = build
                 .put("count", "2")
                 .put("page", "1")
                 .put("name", "李白")
-                .builder();
+                .build();
         HttpHeader httpHeader = HttpHeader.builder();
         httpHeader.add("Accept", "*/*");
         //测试为空,为null
@@ -100,12 +100,12 @@ public class HttpClientSyncImplTest {
         //请求头 参数判断
         HttpClientSyncImpl sync = new HttpClientSyncImpl();
         sync.setHttpConfig(HttpConfig.builder().encode(true).build());
-        MapBuilder<String, String> build = MapBuilder.build();
+        MapBuilder<String, String> build = MapBuilder.builder();
         Map<String, String> params = build
                 .put("count", "2")
                 .put("page", "1")
                 .put("name", "李白")
-                .builder();
+                .build();
         HttpHeader httpHeader = HttpHeader.builder();
         httpHeader.add("Accept", "*/*");
         //测试为空,为null
@@ -137,12 +137,12 @@ public class HttpClientSyncImplTest {
     @Test
     public void testPost1Test() {
         HttpClientSyncImpl sync = new HttpClientSyncImpl();
-        MapBuilder<String, String> mapBuilder = MapBuilder.build();
+        MapBuilder<String, String> mapBuilder = MapBuilder.builder();
         mapBuilder.put("type", "all");
         mapBuilder.put("page", "1");
         mapBuilder.put("count", "10");
         // https://api.apiopen.top/getJoke
-        String result = sync.post("", mapBuilder.builder());
+        String result = sync.post("", mapBuilder.build());
         Assert.assertNotNull("获取失败", result);
     }
 
@@ -160,14 +160,14 @@ public class HttpClientSyncImplTest {
     @Test
     public void testPost3Test() {
         HttpClientSyncImpl sync = new HttpClientSyncImpl();
-        MapBuilder<String, String> mapBuilder = MapBuilder.build();
+        MapBuilder<String, String> mapBuilder = MapBuilder.builder();
         mapBuilder.put("type", "all");
         mapBuilder.put("page", "1");
         mapBuilder.put("count", "10");
         // https://api.apiopen.top/getJoke
         HttpHeader header = HttpHeader.builder().add("Accept", "*/*");
         sync.setHeader(header);
-        String result = sync.post("", mapBuilder.builder());
+        String result = sync.post("", mapBuilder.build());
         Assert.assertNotNull("获取失败", result);
     }
 }