+++ commons-http/src/test/java/com/hb0730/commons/http/support/okhttp3/OkHttp3AsyncImplTest.java
@@ -60,8 +60,8 @@ public class OkHttp3AsyncImplTest {
     @Test
     public void testGetTest() {
         //有参请求
-        MapBuilder<String, String> build = MapBuilder.build();
-        Map<String, String> params = build.put("count", "2").builder();
+        MapBuilder<String, String> build = MapBuilder.builder();
+        Map<String, String> params = build.put("count", "2").build();
         AsyncHttp async = new OkHttp3AsyncImpl();
         //http://poetry.apiopen.top/getTime
         async.get("", params, new HttpCallback() {
@@ -80,8 +80,8 @@ public class OkHttp3AsyncImplTest {
     @Test
     public void testGet1Test() {
         //多参请求
-        MapBuilder<String, String> build = MapBuilder.build();
-        Map<String, String> params = build.put("count", "2").builder();
+        MapBuilder<String, String> build = MapBuilder.builder();
+        Map<String, String> params = build.put("count", "2").build();
         //http://poetry.apiopen.top/poetryFull?page=1
         AsyncHttp async = new OkHttp3AsyncImpl();
         async.get("", params, new HttpCallback() {
@@ -100,12 +100,12 @@ public class OkHttp3AsyncImplTest {
     @Test
     public void testGet2Test() {
         //请求头 参数判断
-        MapBuilder<String, String> build = MapBuilder.build();
+        MapBuilder<String, String> build = MapBuilder.builder();
         Map<String, String> params = build
                 .put("count", "2")
                 .put("page", "1")
                 .put("name", "李白")
-                .builder();
+                .build();
         HttpHeader httpHeader = HttpHeader.builder();
         httpHeader.add("Accept", "*/*");
         //测试为空,为null
@@ -130,12 +130,12 @@ public class OkHttp3AsyncImplTest {
     @Test
     public void testGet3Test() {
         //请求头 参数判断
-        MapBuilder<String, String> build = MapBuilder.build();
+        MapBuilder<String, String> build = MapBuilder.builder();
         Map<String, String> params = build
                 .put("count", "2")
                 .put("page", "1")
                 .put("name", "李白")
-                .builder();
+                .build();
         HttpHeader httpHeader = HttpHeader.builder();
         httpHeader.add("Accept", "*/*");
         //测试为空,为null
@@ -195,13 +195,13 @@ public class OkHttp3AsyncImplTest {
 
     @Test
     public void testPost1Test() {
-        MapBuilder<String, String> mapBuilder = MapBuilder.build();
+        MapBuilder<String, String> mapBuilder = MapBuilder.builder();
         mapBuilder.put("type", "all");
         mapBuilder.put("page", "1");
         mapBuilder.put("count", "10");
         // https://api.apiopen.top/getJoke
         AbstractAsyncHttp async = new OkHttp3AsyncImpl();
-        async.post("", mapBuilder.builder(), new HttpCallback() {
+        async.post("", mapBuilder.build(), new HttpCallback() {
             @Override
             public void success(String result) throws IOException {
                 Assert.assertNotNull("获取失败", result);
@@ -236,7 +236,7 @@ public class OkHttp3AsyncImplTest {
 
     @Test
     public void testPost3Test() {
-        MapBuilder<String, String> mapBuilder = MapBuilder.build();
+        MapBuilder<String, String> mapBuilder = MapBuilder.builder();
         mapBuilder.put("type", "all");
         mapBuilder.put("page", "1");
         mapBuilder.put("count", "10");
@@ -244,7 +244,7 @@ public class OkHttp3AsyncImplTest {
         HttpHeader header = HttpHeader.builder().add("Accept", "*/*");
         AbstractAsyncHttp async = new OkHttp3AsyncImpl();
         async.setHeader(header);
-        async.post("", mapBuilder.builder(), new HttpCallback() {
+        async.post("", mapBuilder.build(), new HttpCallback() {
             @Override
             public void success(String result) throws IOException {
                 Assert.assertNotNull("获取失败", result);